Deduplicate pattern DOM helpers
