+++ php-transformer/src/HtmlToBlocks/Patterns/PatternDomHelpersTrait.php
@@ -1,0 +1,41 @@
+<?php
+declare(strict_types=1);
+
+namespace Automattic\BlocksEngine\PhpTransformer\HtmlToBlocks\Patterns;
+
+use DOMElement;
+
+trait PatternDomHelpersTrait
+{
+    private function attr(DOMElement $element, string $name): string
+    {
+        return $element->hasAttribute($name) ? $element->getAttribute($name) : '';
+    }
+
+    private function hasClass(DOMElement $element, string $className): bool
+    {
+        return in_array($className, preg_split('/\s+/', trim($this->attr($element, 'class'))) ?: array(), true);
+    }
+
+    private function firstChildElement(DOMElement $element, string $tagName): ?DOMElement
+    {
+        foreach ( $element->childNodes as $child ) {
+            if ( $child instanceof DOMElement && strtolower($child->tagName) === $tagName ) {
+                return $child;
+            }
+        }
+
+        return null;
+    }
+
+    private function hasDirectChildElement(DOMElement $element, string $tagName): bool
+    {
+        foreach ( $element->childNodes as $child ) {
+            if ( $child instanceof DOMElement && $tagName === strtolower($child->tagName) ) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+}
